Crate solana_core

The solana library implements the Solana high-performance blockchain architecture. It includes a full Rust implementation of the architecture (see Validator) as well as hooks to GPU implementations of its most paralellizable components (i.e. SigVerify). It also includes command-line tools to spin up validators and a Rust library

Modules

accounts_hash_verifier

admin_rpc_post_init

ancestor_hashes_service

banking_stage

The banking_stage processes Transaction messages. It is intended to be used to construct a software pipeline. The stage uses all available CPU cores and can do its processing in parallel with signature verification on the GPU.

banking_trace

broadcast_stage

A stage to broadcast data from a leader node to validators

cache_block_meta_service

cluster_info_vote_listener

cluster_nodes

cluster_slot_state_verifier

cluster_slots

cluster_slots_service

commitment_service

completed_data_sets_service

CompletedDataSetsService is a hub, that runs different operations when a “completed data set”, also known as a Vec<Entry>, is received by the validator.

consensus

cost_update_service

this service asynchronously reports CostTracker stats

drop_bank_service

duplicate_repair_status

fetch_stage

The fetch_stage batches input from a UDP socket and sends it to a channel.

fork_choice

forward_packet_batches_by_accounts

gen_keys

The gen_keys module makes lots of keypairs

heaviest_subtree_fork_choice

immutable_deserialized_packet

latest_unprocessed_votes 🔒

latest_validator_votes_for_frozen_banks

leader_slot_banking_stage_metrics

leader_slot_banking_stage_timing_metrics

ledger_cleanup_service

The ledger_cleanup_service drops older ledger data to limit disk space usage. The service works by counting the number of live data shreds in the ledger; this can be done quickly and should have a fairly stable correlation to actual bytes. Once the shred count (and thus roughly the byte count) reaches a threshold, the services begins removing data in FIFO order.

ledger_metric_report_service

The ledger_metric_report_service periodically reports ledger store metrics.

multi_iterator_scanner

Provides an iterator interface that create non-conflicting batches of elements to process.

next_leader

optimistic_confirmation_verifier

outstanding_requests

packet_deserializer

Deserializes packets from sigverify stage. Owned by banking stage.

packet_threshold

poh_timing_report_service

PohTimingReportService module

poh_timing_reporter

A poh_timing_reporter module implement poh timing point and timing reporter structs.

progress_map

qos_service

Quality of service for block producer. Provides logic and functions to allow a Leader to prioritize how transactions are included in blocks, and optimize those blocks.

read_write_account_set

repair_generic_traversal

repair_response

repair_service

The repair_service module implements the tools necessary to generate a thread which regularly finds missing shreds in the ledger and sends repair requests for those shreds

repair_weight

repair_weighted_traversal

replay_stage

The replay_stage replays transactions broadcast by the leader.

request_response

result 🔒

The result module exposes a Result type that propagates one of many different Error types.

retransmit_stage

The retransmit_stage retransmits shreds between validators

rewards_recorder_service

sample_performance_service

serve_repair

serve_repair_service

shred_fetch_stage 🔒

The shred_fetch_stage pulls shreds from UDP sockets and sends it to a channel.

sigverify

The sigverify module provides digital signature verification functions. By default, signatures are verified in parallel using all available CPU cores. When perf-libs are available signature verification is offloaded to the GPU.

sigverify_shreds

sigverify_stage

The sigverify_stage implements the signature verification stage of the TPU. It receives a list of lists of packets and outputs the same list, but tags each top-level list with a list of booleans, telling the next stage whether the signature in that packet is valid. It assumes each packet contains one transaction. All processing is done on the CPU by default and on a GPU if perf-libs are available

snapshot_packager_service

staked_nodes_updater_service

stats_reporter_service

system_monitor_service

tower1_7_14 🔒

tower_storage

tpu

The tpu module implements the Transaction Processing Unit, a multi-stage transaction processing pipeline in software.

tracer_packet_stats

tree_diff

tvu

The tvu module implements the Transaction Validation Unit, a multi-stage transaction validation pipeline in software.

unfrozen_gossip_verified_vote_hashes

unprocessed_packet_batches

unprocessed_transaction_storage

validator

The validator module hosts all the validator microservices.

verified_vote_packets

vote_simulator

vote_stake_tracker

voting_service

warm_quic_cache_service

window_service

window_service handles the data plane incoming shreds, storing them in blockstore and retransmitting where required