Struct solana_core::ancestor_hashes_service::AncestorHashesService

pub struct AncestorHashesService {
    thread_hdls: Vec<JoinHandle<()>>,
}

Fields
thread_hdls: Vec<JoinHandle<()>>
Implementations

impl AncestorHashesService

pub fn new( exit: Arc<AtomicBool>, blockstore: Arc<Blockstore>, ancestor_hashes_request_socket: Arc<UdpSocket>, repair_info: RepairInfo, ancestor_hashes_replay_update_receiver: AncestorHashesReplayUpdateReceiver ) -> Self

pub fn join(self) -> Result<()>

fn run_responses_listener( ancestor_hashes_request_statuses: Arc<DashMap<Slot, AncestorRequestStatus>>, response_receiver: PacketBatchReceiver, blockstore: Arc<Blockstore>, outstanding_requests: Arc<RwLock<OutstandingRequests<AncestorHashesRepairType>>>, exit: Arc<AtomicBool>, duplicate_slots_reset_sender: DuplicateSlotsResetSender, retryable_slots_sender: Sender<Slot>, cluster_info: Arc<ClusterInfo>, ancestor_socket: Arc<UdpSocket> ) -> JoinHandle<()>

Listen for responses to our ancestors hashes repair requests

fn process_new_packets_from_channel( ancestor_hashes_request_statuses: &DashMap<Slot, AncestorRequestStatus>, response_receiver: &PacketBatchReceiver, blockstore: &Blockstore, outstanding_requests: &RwLock<OutstandingRequests<AncestorHashesRepairType>>, stats: &mut AncestorHashesResponsesStats, packet_threshold: &mut DynamicPacketToProcessThreshold, duplicate_slots_reset_sender: &DuplicateSlotsResetSender, retryable_slots_sender: &Sender<Slot>, keypair: &Keypair, ancestor_socket: &UdpSocket ) -> Result<(), Error>

Process messages from the network

fn process_packet_batch( ancestor_hashes_request_statuses: &DashMap<Slot, AncestorRequestStatus>, packet_batch: PacketBatch, stats: &mut AncestorHashesResponsesStats, outstanding_requests: &RwLock<OutstandingRequests<AncestorHashesRepairType>>, blockstore: &Blockstore, duplicate_slots_reset_sender: &DuplicateSlotsResetSender, retryable_slots_sender: &Sender<Slot>, keypair: &Keypair, ancestor_socket: &UdpSocket )

fn verify_and_process_ancestor_response( packet: &Packet, ancestor_hashes_request_statuses: &DashMap<Slot, AncestorRequestStatus>, stats: &mut AncestorHashesResponsesStats, outstanding_requests: &RwLock<OutstandingRequests<AncestorHashesRepairType>>, blockstore: &Blockstore, keypair: &Keypair, ancestor_socket: &UdpSocket ) -> Option<(Slot, DuplicateAncestorDecision)>

Returns Some((request_slot, decision)), where decision is an actionable result after processing sufficient responses for the subject of the query, request_slot

fn handle_ancestor_request_decision( slot: Slot, decision: DuplicateAncestorDecision, duplicate_slots_reset_sender: &DuplicateSlotsResetSender, retryable_slots_sender: &Sender<Slot> )

fn process_replay_updates( ancestor_hashes_replay_update_receiver: &AncestorHashesReplayUpdateReceiver, ancestor_hashes_request_statuses: &DashMap<Slot, AncestorRequestStatus>, dead_slot_pool: &mut HashSet<Slot>, repairable_dead_slot_pool: &mut HashSet<Slot>, root_slot: Slot )

fn run_manage_ancestor_requests( ancestor_hashes_request_statuses: Arc<DashMap<Slot, AncestorRequestStatus>>, ancestor_hashes_request_socket: Arc<UdpSocket>, repair_info: RepairInfo, outstanding_requests: Arc<RwLock<OutstandingRequests<AncestorHashesRepairType>>>, exit: Arc<AtomicBool>, ancestor_hashes_replay_update_receiver: AncestorHashesReplayUpdateReceiver, retryable_slots_receiver: Receiver<Slot> ) -> JoinHandle<()>

fn manage_ancestor_requests( ancestor_hashes_request_statuses: &DashMap<Slot, AncestorRequestStatus>, ancestor_hashes_request_socket: &UdpSocket, repair_info: &RepairInfo, outstanding_requests: &RwLock<OutstandingRequests<AncestorHashesRepairType>>, ancestor_hashes_replay_update_receiver: &AncestorHashesReplayUpdateReceiver, retryable_slots_receiver: &Receiver<Slot>, serve_repair: &ServeRepair, repair_stats: &mut AncestorRepairRequestsStats, dead_slot_pool: &mut HashSet<Slot>, repairable_dead_slot_pool: &mut HashSet<Slot>, request_throttle: &mut Vec<u64> )

fn find_epoch_slots_frozen_dead_slots( cluster_slots: &ClusterSlots, dead_slot_pool: &mut HashSet<Slot>, repairable_dead_slot_pool: &mut HashSet<Slot>, root_bank: &Bank )

Find if any dead slots in dead_slot_pool have been frozen by sufficient number of nodes in the cluster to justify adding to the repairable_dead_slot_pool.

fn initiate_ancestor_hashes_requests_for_duplicate_slot( ancestor_hashes_request_statuses: &DashMap<Slot, AncestorRequestStatus>, ancestor_hashes_request_socket: &UdpSocket, cluster_slots: &ClusterSlots, serve_repair: &ServeRepair, repair_validators: &Option<HashSet<Pubkey>>, duplicate_slot: Slot, repair_stats: &mut AncestorRepairRequestsStats, outstanding_requests: &RwLock<OutstandingRequests<AncestorHashesRepairType>>, identity_keypair: &Keypair ) -> bool

Returns true if a request was successfully made and the status added to ancestor_hashes_request_statuses