Module solana_core::broadcast_stage

A stage to broadcast data from a leader node to validators

Modules

broadcast_duplicates_run

broadcast_fake_shreds_run ğŸ”’

broadcast_metrics

broadcast_utils ğŸ”’

fail_entry_verification_broadcast_run ğŸ”’

standard_broadcast_run ğŸ”’

Structs

BroadcastStage

Finalizer ğŸ”’

Enums

BroadcastStageReturnType

BroadcastStageType

Constants

CLUSTER_NODES_CACHE_NUM_EPOCH_CAP ğŸ”’

CLUSTER_NODES_CACHE_TTL ğŸ”’

NUM_INSERT_THREADS ğŸ”’

RECV_TIMEOUT ğŸ”’

Statics

INC_NEW_COUNTER ğŸ”’

INIT_HOOK ğŸ”’

Traits

BroadcastRun ğŸ”’

Functions

broadcast_shreds

Broadcasts shreds from the leader (i.e. this node) to the root of the turbine retransmit tree for each shred.

update_peer_stats ğŸ”’

Type Definitions

RecordReceiver ğŸ”’

RetransmitSlotsReceiver ğŸ”’

RetransmitSlotsSender ğŸ”’

TransmitReceiver ğŸ”’