Struct solana_core::broadcast_stage::BroadcastStage

pub struct BroadcastStage {
    thread_hdls: Vec<JoinHandle<BroadcastStageReturnType>>,
}

Fields
thread_hdls: Vec<JoinHandle<BroadcastStageReturnType>>
Implementations

impl BroadcastStage

fn run( cluster_info: Arc<ClusterInfo>, blockstore: &Blockstore, receiver: &Receiver<WorkingBankEntry>, socket_sender: &Sender<(Arc<Vec<Shred>>, Option<BroadcastShredBatchInfo>)>, blockstore_sender: &Sender<(Arc<Vec<Shred>>, Option<BroadcastShredBatchInfo>)>, broadcast_stage_run: impl BroadcastRun ) -> BroadcastStageReturnType

fn handle_error( r: Result<(), Error>, name: &str ) -> Option<BroadcastStageReturnType>

fn new( socks: Vec<UdpSocket>, cluster_info: Arc<ClusterInfo>, receiver: Receiver<WorkingBankEntry>, retransmit_slots_receiver: Receiver<Slot>, exit: Arc<AtomicBool>, blockstore: Arc<Blockstore>, bank_forks: Arc<RwLock<BankForks>>, broadcast_stage_run: impl BroadcastRun + Send + 'static + Clone ) -> Self

Service to broadcast messages from the leader to layer 1 nodes. See cluster_info for network layer definitions.

Arguments

sock - Socket to send from.

exit - Boolean to signal system exit.

cluster_info - ClusterInfo structure

window - Cache of Shreds that we have broadcast

receiver - Receive channel for Shreds to be retransmitted to all the layer 1 nodes.

exit_sender - Set to true when this service exits, allows rest of Tpu to exit cleanly. Otherwise, when a Tpu closes, it only closes the stages that come after it. The stages that come before could be blocked on a receive, and never notice that they need to exit. Now, if any stage of the Tpu closes, it will lead to closing the WriteStage (b/c WriteStage is the last stage in the pipeline), which will then close Broadcast service, which will then close FetchStage in the Tpu, and then the rest of the Tpu, completing the cycle.

fn check_retransmit_signals( blockstore: &Blockstore, retransmit_slots_receiver: &Receiver<Slot>, socket_sender: &Sender<(Arc<Vec<Shred>>, Option<BroadcastShredBatchInfo>)> ) -> Result<(), Error>

pub fn join(self) -> Result<BroadcastStageReturnType>