Struct solana_core::heaviest_subtree_fork_choice::HeaviestSubtreeForkChoice

pub struct HeaviestSubtreeForkChoice {
    fork_infos: HashMap<SlotHashKey, ForkInfo>,
    latest_votes: HashMap<Pubkey, SlotHashKey>,
    tree_root: SlotHashKey,
    last_root_time: Instant,
}

Fields
fork_infos: HashMap<SlotHashKey, ForkInfo>latest_votes: HashMap<Pubkey, SlotHashKey>tree_root: SlotHashKeylast_root_time: Instant
Implementations

impl HeaviestSubtreeForkChoice

pub fn new(tree_root: SlotHashKey) -> Self

pub fn new_from_frozen_banks( root: SlotHashKey, frozen_banks: &[Arc<Bank>] ) -> Self

pub fn new_from_bank_forks(bank_forks: &BankForks) -> Self

pub fn contains_block(&self, key: &SlotHashKey) -> bool

pub fn best_slot(&self, key: &SlotHashKey) -> Option<SlotHashKey>

pub fn best_overall_slot(&self) -> SlotHashKey

pub fn stake_voted_subtree(&self, key: &SlotHashKey) -> Option<u64>

pub fn tree_root(&self) -> SlotHashKey

pub fn max_by_weight(&self, slot1: SlotHashKey, slot2: SlotHashKey) -> Ordering

pub fn add_votes<'a, 'b>( &'a mut self, pubkey_votes: impl Iterator<Item = impl Borrow<(Pubkey, SlotHashKey)> + 'b>, epoch_stakes: &HashMap<Epoch, EpochStakes>, epoch_schedule: &EpochSchedule ) -> SlotHashKey

pub fn is_empty(&self) -> bool

pub fn set_tree_root(&mut self, new_root: SlotHashKey)

pub fn purge_prune( &mut self, new_root: SlotHashKey ) -> (Vec<SlotHashKey>, Vec<Self>)

Purges all slots < new_root and prunes subtrees with slots > new_root not descending from new_root. Also if the resulting tree is non-empty, updates self.tree_root to new_root Returns (purged slots, pruned subtrees)

pub fn add_root_parent(&mut self, root_parent: SlotHashKey)

pub fn add_new_leaf_slot( &mut self, slot_hash_key: SlotHashKey, parent: Option<SlotHashKey> )

fn is_best_child(&self, maybe_best_child: &SlotHashKey) -> bool

pub fn all_slots_stake_voted_subtree( &self ) -> impl Iterator<Item = (&SlotHashKey, u64)>

pub fn split_off(&mut self, slot_hash_key: &SlotHashKey) -> Self

Split off the node at slot_hash_key and propagate the stake subtraction up to the root of the tree.

Assumes that slot_hash_key is not the tree_root Returns the subtree originating from slot_hash_key

pub fn merge( &mut self, other: HeaviestSubtreeForkChoice, merge_leaf: &SlotHashKey, epoch_stakes: &HashMap<Epoch, EpochStakes>, epoch_schedule: &EpochSchedule )

pub fn stake_voted_at(&self, slot: &SlotHashKey) -> Option<u64>

pub fn latest_invalid_ancestor( &self, slot_hash_key: &SlotHashKey ) -> Option<Slot>

pub fn is_duplicate_confirmed( &self, slot_hash_key: &SlotHashKey ) -> Option<bool>

pub fn is_unconfirmed_duplicate( &self, slot_hash_key: &SlotHashKey ) -> Option<bool>

Returns if the exact node with the specified key has been explicitly marked as a duplicate slot (doesn’t count ancestors being marked as duplicate).

pub fn is_candidate(&self, slot_hash_key: &SlotHashKey) -> Option<bool>

Returns false if the node or any of its ancestors have been marked as duplicate

pub fn is_strict_ancestor( &self, maybe_ancestor_key: &SlotHashKey, node_key: &SlotHashKey ) -> bool

Returns if a node with slot maybe_ancestor_slot is an ancestor of the node with key node_key

fn propagate_new_leaf( &mut self, slot_hash_key: &SlotHashKey, parent_slot_hash_key: &SlotHashKey )

fn insert_aggregate_operations( &self, update_operations: &mut BTreeMap<(SlotHashKey, UpdateLabel), UpdateOperation>, slot_hash_key: SlotHashKey )

fn do_insert_aggregate_operations_across_ancestors( &self, update_operations: &mut BTreeMap<(SlotHashKey, UpdateLabel), UpdateOperation>, modify_fork_validity: Option<UpdateOperation>, slot_hash_key: SlotHashKey )

fn do_insert_aggregate_operation( &self, update_operations: &mut BTreeMap<(SlotHashKey, UpdateLabel), UpdateOperation>, modify_fork_validity: &Option<UpdateOperation>, slot_hash_key: SlotHashKey ) -> bool

fn ancestor_iterator( &self, start_slot_hash_key: SlotHashKey ) -> AncestorIterator<'_>

fn aggregate_slot(&mut self, slot_hash_key: SlotHashKey)

fn mark_fork_valid(&mut self, fork_to_modify_key: SlotHashKey, valid_slot: Slot)

Mark that valid_slot on the fork starting at fork_to_modify has been marked valid. Note we don’t need the hash for valid_slot because slot number uniquely identifies a node on a single fork.

fn mark_fork_invalid( &mut self, fork_to_modify_key: SlotHashKey, invalid_slot: Slot )

Mark that invalid_slot on the fork starting at fork_to_modify has been marked invalid. Note we don’t need the hash for invalid_slot because slot number uniquely identifies a node on a single fork.

fn generate_update_operations<'a, 'b>( &'a mut self, pubkey_votes: impl Iterator<Item = impl Borrow<(Pubkey, SlotHashKey)> + 'b>, epoch_stakes: &HashMap<Epoch, EpochStakes>, epoch_schedule: &EpochSchedule ) -> BTreeMap<(SlotHashKey, UpdateLabel), UpdateOperation>

fn process_update_operations( &mut self, update_operations: BTreeMap<(SlotHashKey, UpdateLabel), UpdateOperation> )

fn add_slot_stake(&mut self, slot_hash_key: &SlotHashKey, stake: u64)

fn subtract_slot_stake(&mut self, slot_hash_key: &SlotHashKey, stake: u64)

fn parent(&self, slot_hash_key: &SlotHashKey) -> Option<SlotHashKey>

fn print_state(&self)

fn heaviest_slot_on_same_voted_fork(&self, tower: &Tower) -> Option<SlotHashKey>