Struct solana_core::ledger_cleanup_service::LedgerCleanupService

pub struct LedgerCleanupService {
    t_cleanup: JoinHandle<()>,
}

Fields
t_cleanup: JoinHandle<()>
Implementations

impl LedgerCleanupService

pub fn new( new_root_receiver: Receiver<Slot>, blockstore: Arc<Blockstore>, max_ledger_shreds: u64, exit: &Arc<AtomicBool> ) -> Self

fn find_slots_to_clean( blockstore: &Arc<Blockstore>, root: Slot, max_ledger_shreds: u64 ) -> (bool, Slot, u64)

A helper function to cleanup_ledger which returns a tuple of the following four elements suggesting whether to clean up the ledger:

Return value (bool, Slot, u64):

slots_to_clean (bool): a boolean value indicating whether there are any slots to clean. If true, then cleanup_ledger function will then proceed with the ledger cleanup.

lowest_slot_to_purge (Slot): the lowest slot to purge. Any slot which is older or equal to lowest_slot_to_purge will be cleaned up.

total_shreds (u64): the total estimated number of shreds before the root.

fn receive_new_roots( new_root_receiver: &Receiver<Slot> ) -> Result<Slot, RecvTimeoutError>

pub fn cleanup_ledger( new_root_receiver: &Receiver<Slot>, blockstore: &Arc<Blockstore>, max_ledger_shreds: u64, last_purge_slot: &mut u64, purge_interval: u64 ) -> Result<(), RecvTimeoutError>

Checks for new roots and initiates a cleanup if the last cleanup was at least purge_interval slots ago. A cleanup will no-op if the ledger already has fewer than max_ledger_shreds; otherwise, the cleanup will purge enough slots to get the ledger size below max_ledger_shreds.

Arguments

new_root_receiver: signal receiver which contains the information about what Slot is the current root.

max_ledger_shreds: the number of shreds to keep since the new root.

last_purge_slot: an both an input and output parameter indicating the id of the last purged slot. As an input parameter, it works together with purge_interval on whether it is too early to perform ledger cleanup. As an output parameter, it will be updated if this function actually performs the ledger cleanup.

purge_interval: the minimum slot interval between two ledger cleanup. When the root derived from new_root_receiver minus last_purge_slot is fewer than purge_interval, the function will simply return Ok without actually running the ledger cleanup. In this case, purge_interval will remain unchanged.

Also see blockstore::purge_slot.

fn report_disk_metrics( pre: BlockstoreResult<u64>, post: BlockstoreResult<u64>, total_shreds: u64 )

pub fn join(self) -> Result<()>