Module solana_core::repair_generic_traversal

Structs

GenericTraversal 🔒

Functions

get_closest_completion

Finds repairs for slots that are closest to completion (# of missing shreds). Additionaly we repair up to their oldest full ancestor (using blockstore fork info).

get_unknown_last_index

Does a generic traversal and inserts all slots that have a missing last index prioritized by how many shreds have been received

get_unrepaired_path 🔒

Path of broken parents from start_slot to earliest ancestor not yet seen Uses blockstore for fork information