Struct solana_core::repair_weight::RepairWeight

pub struct RepairWeight {
    trees: HashMap<Slot, HeaviestSubtreeForkChoice>,
    pruned_trees: HashMap<Slot, HeaviestSubtreeForkChoice>,
    slot_to_tree: HashMap<Slot, TreeRoot>,
    root: Slot,
}

Fields
trees: HashMap<Slot, HeaviestSubtreeForkChoice>pruned_trees: HashMap<Slot, HeaviestSubtreeForkChoice>slot_to_tree: HashMap<Slot, TreeRoot>root: Slot
Implementations

impl RepairWeight

pub fn new(root: Slot) -> Self

pub fn add_votes<I>( &mut self, blockstore: &Blockstore, votes: I, epoch_stakes: &HashMap<Epoch, EpochStakes>, epoch_schedule: &EpochSchedule )where I: Iterator<Item = (Slot, Vec<Pubkey>)>,

pub fn get_best_weighted_repairs( &mut self, blockstore: &Blockstore, epoch_stakes: &HashMap<Epoch, EpochStakes>, epoch_schedule: &EpochSchedule, max_new_orphans: usize, max_new_shreds: usize, max_unknown_last_index_repairs: usize, max_closest_completion_repairs: usize, repair_timing: &mut RepairTiming, stats: &mut BestRepairsStats ) -> Vec<ShredRepairType>

pub fn split_off(&mut self, slot: Slot)

Split slot and descendants into an orphan tree in repair weighting.

If repair holds a subtree ST which contains slot, we split ST into (T, T') where T' is a subtree rooted at slot and T is what remains (if anything) of ST after the split. If T is non-empty, it is inserted back into the set which held the original ST (self.trees for potentially rootable trees, or self.pruned_trees for pruned trees). T' is always inserted into the potentially rootable set self.trees. This function removes and destroys the original ST.

Assumes that slot is greater than self.root.

pub fn set_root(&mut self, new_root: Slot)

pub fn root(&self) -> Slot

fn get_best_shreds( &mut self, blockstore: &Blockstore, slot_meta_cache: &mut HashMap<Slot, Option<SlotMeta>>, repairs: &mut Vec<ShredRepairType>, max_new_shreds: usize )

fn get_best_orphans( &mut self, blockstore: &Blockstore, processed_slots: &mut HashSet<Slot>, repairs: &mut Vec<ShredRepairType>, epoch_stakes: &HashMap<Epoch, EpochStakes>, epoch_schedule: &EpochSchedule, max_new_orphans: usize )

fn get_best_unknown_last_index( &mut self, blockstore: &Blockstore, slot_meta_cache: &mut HashMap<Slot, Option<SlotMeta>>, processed_slots: &mut HashSet<Slot>, max_new_repairs: usize ) -> Vec<ShredRepairType>

For all remaining trees (orphan and rooted), generate repairs for slots missing last_index info prioritized by # shreds received.

fn get_best_closest_completion( &mut self, blockstore: &Blockstore, slot_meta_cache: &mut HashMap<Slot, Option<SlotMeta>>, processed_slots: &mut HashSet<Slot>, max_new_repairs: usize ) -> (Vec<ShredRepairType>, usize)

For all remaining trees (orphan and rooted), generate repairs for subtrees that have last index info but are missing shreds prioritized by how close to completion they are. These repairs are also prioritized by age of ancestors, so slots close to completion will first start by repairing broken ancestors.

fn update_orphan_ancestors( &mut self, blockstore: &Blockstore, orphan_tree_root: Slot, epoch_stakes: &HashMap<Epoch, EpochStakes>, epoch_schedule: &EpochSchedule ) -> Option<Slot>

Attempts to chain the orphan subtree rooted at orphan_tree_root to any earlier subtree with new ancestry information in blockstore. Returns the earliest known ancestor of heaviest_tree_root.

fn get_tree( &mut self, tree_root: &TreeRoot ) -> Option<&mut HeaviestSubtreeForkChoice>

fn remove_tree( &mut self, tree_root: &TreeRoot ) -> Option<HeaviestSubtreeForkChoice>

fn get_tree_root(&self, slot: Slot) -> Option<TreeRoot>

fn insert_new_tree(&mut self, new_tree_root: Slot)

Assumes that new_tree_root does not already exist in self.trees

fn insert_new_pruned_tree(&mut self, new_pruned_tree_root: Slot)

Assumes that new_pruned_tree_root does not already exist in self.pruned_trees

fn find_ancestor_subtree_of_slot( &self, blockstore: &Blockstore, slot: Slot ) -> (VecDeque<Slot>, Option<TreeRoot>)

Finds any ancestors avaiable from blockstore for slot. Ancestor search is stopped when finding one that chains to any tree in self.trees or self.pruned_trees or if the ancestor is < self.root.

Returns ancestors (including the one that triggered the search stop condition), and possibly the tree_root that slot belongs to

fn merge_trees( &mut self, root1: TreeRoot, root2: TreeRoot, merge_leaf: Slot, epoch_stakes: &HashMap<Epoch, EpochStakes>, epoch_schedule: &EpochSchedule )

Attaches tree1 rooted at root1 to tree2 rooted at root2 at the leaf in tree2 given by merge_leaf Expects root1 and root2 to exist in self.trees or self.pruned_trees

fn rename_tree_root( &mut self, tree1: &HeaviestSubtreeForkChoice, root2: TreeRoot )

fn sort_by_stake_weight_slot(slot_stake_voted: &mut [(Slot, u64)])