Struct solana_core::replay_stage::ReplayStage

pub struct ReplayStage {
    t_replay: JoinHandle<()>,
    commitment_service: AggregateCommitmentService,
}

Fields
t_replay: JoinHandle<()>commitment_service: AggregateCommitmentService
Implementations

impl ReplayStage

pub fn new( config: ReplayStageConfig, blockstore: Arc<Blockstore>, bank_forks: Arc<RwLock<BankForks>>, cluster_info: Arc<ClusterInfo>, ledger_signal_receiver: Receiver<bool>, duplicate_slots_receiver: Receiver<Slot>, poh_recorder: Arc<RwLock<PohRecorder>>, maybe_process_blockstore: Option<ProcessBlockStore<'_>>, vote_tracker: Arc<VoteTracker>, cluster_slots: Arc<ClusterSlots>, retransmit_slots_sender: Sender<Slot>, epoch_slots_frozen_receiver: DuplicateSlotsResetReceiver, replay_vote_sender: ReplayVoteSender, gossip_duplicate_confirmed_slots_receiver: GossipDuplicateConfirmedSlotsReceiver, gossip_verified_vote_hash_receiver: GossipVerifiedVoteHashReceiver, cluster_slots_update_sender: ClusterSlotsUpdateSender, cost_update_sender: Sender<CostUpdate>, voting_sender: Sender<VoteOp>, drop_bank_sender: Sender<Vec<Arc<Bank>>>, block_metadata_notifier: Option<BlockMetadataNotifierLock>, log_messages_bytes_limit: Option<usize>, prioritization_fee_cache: Arc<PrioritizationFeeCache>, dumped_slots_sender: DumpedSlotsSender, banking_tracer: Arc<BankingTracer> ) -> Result<Self, String>

fn check_for_vote_only_mode( heaviest_bank_slot: Slot, forks_root: Slot, in_vote_only_mode: &AtomicBool, bank_forks: &RwLock<BankForks> )

fn maybe_retransmit_unpropagated_slots( metric_name: &'static str, retransmit_slots_sender: &Sender<Slot>, progress: &mut ProgressMap, latest_leader_slot: Slot )

fn retransmit_latest_unpropagated_leader_slot( poh_recorder: &Arc<RwLock<PohRecorder>>, retransmit_slots_sender: &Sender<Slot>, progress: &mut ProgressMap )

fn is_partition_detected( ancestors: &HashMap<Slot, HashSet<Slot>>, last_voted_slot: Slot, heaviest_slot: Slot ) -> bool

fn initialize_progress_and_fork_choice_with_locked_bank_forks( bank_forks: &RwLock<BankForks>, my_pubkey: &Pubkey, vote_account: &Pubkey ) -> (ProgressMap, HeaviestSubtreeForkChoice)

pub fn initialize_progress_and_fork_choice( root_bank: &Bank, frozen_banks: Vec<Arc<Bank>>, my_pubkey: &Pubkey, vote_account: &Pubkey ) -> (ProgressMap, HeaviestSubtreeForkChoice)

pub fn dump_then_repair_correct_slots( duplicate_slots_to_repair: &mut HashMap<Slot, Hash>, ancestors: &mut HashMap<Slot, HashSet<Slot>>, descendants: &mut HashMap<Slot, HashSet<Slot>>, progress: &mut ProgressMap, bank_forks: &RwLock<BankForks>, blockstore: &Blockstore, poh_bank_slot: Option<Slot>, purge_repair_slot_counter: &mut BTreeMap<Slot, usize>, dumped_slots_sender: &DumpedSlotsSender, my_pubkey: &Pubkey, leader_schedule_cache: &LeaderScheduleCache )

fn process_epoch_slots_frozen_dead_slots( pubkey: &Pubkey, blockstore: &Blockstore, epoch_slots_frozen_receiver: &DuplicateSlotsResetReceiver, duplicate_slots_tracker: &mut BTreeSet<Slot>, gossip_duplicate_confirmed_slots: &BTreeMap<Slot, Hash>, epoch_slots_frozen_slots: &mut BTreeMap<Slot, Hash>, progress: &mut ProgressMap, fork_choice: &mut HeaviestSubtreeForkChoice, bank_forks: &RwLock<BankForks>, duplicate_slots_to_repair: &mut HashMap<Slot, Hash>, ancestor_hashes_replay_update_sender: &AncestorHashesReplayUpdateSender, purge_repair_slot_counter: &mut BTreeMap<Slot, usize> )

fn purge_unconfirmed_duplicate_slot( duplicate_slot: Slot, ancestors: &mut HashMap<Slot, HashSet<Slot>>, descendants: &mut HashMap<Slot, HashSet<Slot>>, progress: &mut ProgressMap, root_bank: &Bank, bank_forks: &RwLock<BankForks>, blockstore: &Blockstore )

fn purge_ancestors_descendants( slot: Slot, slot_descendants: &HashSet<Slot>, ancestors: &mut HashMap<Slot, HashSet<Slot>>, descendants: &mut HashMap<Slot, HashSet<Slot>> )

fn process_gossip_duplicate_confirmed_slots( gossip_duplicate_confirmed_slots_receiver: &GossipDuplicateConfirmedSlotsReceiver, blockstore: &Blockstore, duplicate_slots_tracker: &mut BTreeSet<Slot>, gossip_duplicate_confirmed_slots: &mut BTreeMap<Slot, Hash>, epoch_slots_frozen_slots: &mut BTreeMap<Slot, Hash>, bank_forks: &RwLock<BankForks>, progress: &mut ProgressMap, fork_choice: &mut HeaviestSubtreeForkChoice, duplicate_slots_to_repair: &mut HashMap<Slot, Hash>, ancestor_hashes_replay_update_sender: &AncestorHashesReplayUpdateSender, purge_repair_slot_counter: &mut BTreeMap<Slot, usize> )

fn process_gossip_verified_vote_hashes( gossip_verified_vote_hash_receiver: &GossipVerifiedVoteHashReceiver, unfrozen_gossip_verified_vote_hashes: &mut UnfrozenGossipVerifiedVoteHashes, heaviest_subtree_fork_choice: &HeaviestSubtreeForkChoice, latest_validator_votes_for_frozen_banks: &mut LatestValidatorVotesForFrozenBanks )

fn process_duplicate_slots( blockstore: &Blockstore, duplicate_slots_receiver: &Receiver<Slot>, duplicate_slots_tracker: &mut BTreeSet<Slot>, gossip_duplicate_confirmed_slots: &BTreeMap<Slot, Hash>, epoch_slots_frozen_slots: &mut BTreeMap<Slot, Hash>, bank_forks: &RwLock<BankForks>, progress: &mut ProgressMap, fork_choice: &mut HeaviestSubtreeForkChoice, duplicate_slots_to_repair: &mut HashMap<Slot, Hash>, ancestor_hashes_replay_update_sender: &AncestorHashesReplayUpdateSender, purge_repair_slot_counter: &mut BTreeMap<Slot, usize> )

fn log_leader_change( my_pubkey: &Pubkey, bank_slot: Slot, current_leader: &mut Option<Pubkey>, new_leader: &Pubkey )

fn check_propagation_for_start_leader( poh_slot: Slot, parent_slot: Slot, progress_map: &ProgressMap ) -> bool

fn should_retransmit(poh_slot: Slot, last_retransmit_slot: &mut Slot) -> bool

fn maybe_start_leader( my_pubkey: &Pubkey, bank_forks: &Arc<RwLock<BankForks>>, poh_recorder: &Arc<RwLock<PohRecorder>>, leader_schedule_cache: &Arc<LeaderScheduleCache>, rpc_subscriptions: &Arc<RpcSubscriptions>, progress_map: &mut ProgressMap, retransmit_slots_sender: &Sender<Slot>, skipped_slots_info: &mut SkippedSlotsInfo, banking_tracer: &Arc<BankingTracer>, has_new_vote_been_rooted: bool, track_transaction_indexes: bool )

fn replay_blockstore_into_bank( bank: &Arc<Bank>, blockstore: &Blockstore, replay_stats: &RwLock<ReplaySlotStats>, replay_progress: &RwLock<ConfirmationProgress>, transaction_status_sender: Option<&TransactionStatusSender>, replay_vote_sender: &ReplayVoteSender, verify_recyclers: &VerifyRecyclers, log_messages_bytes_limit: Option<usize>, prioritization_fee_cache: &PrioritizationFeeCache ) -> Result<usize, BlockstoreProcessorError>

fn mark_dead_slot( blockstore: &Blockstore, bank: &Bank, root: Slot, err: &BlockstoreProcessorError, rpc_subscriptions: &Arc<RpcSubscriptions>, duplicate_slots_tracker: &mut BTreeSet<Slot>, gossip_duplicate_confirmed_slots: &BTreeMap<Slot, Hash>, epoch_slots_frozen_slots: &mut BTreeMap<Slot, Hash>, progress: &mut ProgressMap, heaviest_subtree_fork_choice: &mut HeaviestSubtreeForkChoice, duplicate_slots_to_repair: &mut HashMap<Slot, Hash>, ancestor_hashes_replay_update_sender: &AncestorHashesReplayUpdateSender, purge_repair_slot_counter: &mut BTreeMap<Slot, usize> )

fn handle_votable_bank( bank: &Arc<Bank>, switch_fork_decision: &SwitchForkDecision, bank_forks: &Arc<RwLock<BankForks>>, tower: &mut Tower, progress: &mut ProgressMap, vote_account_pubkey: &Pubkey, identity_keypair: &Keypair, authorized_voter_keypairs: &[Arc<Keypair>], blockstore: &Arc<Blockstore>, leader_schedule_cache: &Arc<LeaderScheduleCache>, lockouts_sender: &Sender<CommitmentAggregationData>, accounts_background_request_sender: &AbsRequestSender, latest_root_senders: &[Sender<Slot>], rpc_subscriptions: &Arc<RpcSubscriptions>, block_commitment_cache: &Arc<RwLock<BlockCommitmentCache>>, heaviest_subtree_fork_choice: &mut HeaviestSubtreeForkChoice, bank_notification_sender: &Option<BankNotificationSender>, duplicate_slots_tracker: &mut BTreeSet<Slot>, gossip_duplicate_confirmed_slots: &mut BTreeMap<Slot, Hash>, unfrozen_gossip_verified_vote_hashes: &mut UnfrozenGossipVerifiedVoteHashes, vote_signatures: &mut Vec<Signature>, has_new_vote_been_rooted: &mut bool, replay_timing: &mut ReplayTiming, voting_sender: &Sender<VoteOp>, epoch_slots_frozen_slots: &mut BTreeMap<Slot, Hash>, drop_bank_sender: &Sender<Vec<Arc<Bank>>>, wait_to_vote_slot: Option<Slot> )

fn generate_vote_tx( node_keypair: &Keypair, bank: &Bank, vote_account_pubkey: &Pubkey, authorized_voter_keypairs: &[Arc<Keypair>], vote: VoteTransaction, switch_fork_decision: &SwitchForkDecision, vote_signatures: &mut Vec<Signature>, has_new_vote_been_rooted: bool, wait_to_vote_slot: Option<Slot> ) -> Option<Transaction>

fn refresh_last_vote( tower: &mut Tower, heaviest_bank_on_same_fork: &Bank, my_latest_landed_vote: Slot, vote_account_pubkey: &Pubkey, identity_keypair: &Keypair, authorized_voter_keypairs: &[Arc<Keypair>], vote_signatures: &mut Vec<Signature>, has_new_vote_been_rooted: bool, last_vote_refresh_time: &mut LastVoteRefreshTime, voting_sender: &Sender<VoteOp>, wait_to_vote_slot: Option<Slot> )

fn push_vote( bank: &Bank, vote_account_pubkey: &Pubkey, identity_keypair: &Keypair, authorized_voter_keypairs: &[Arc<Keypair>], tower: &mut Tower, switch_fork_decision: &SwitchForkDecision, vote_signatures: &mut Vec<Signature>, has_new_vote_been_rooted: bool, replay_timing: &mut ReplayTiming, voting_sender: &Sender<VoteOp>, wait_to_vote_slot: Option<Slot> )

fn update_commitment_cache( bank: Arc<Bank>, root: Slot, total_stake: Stake, lockouts_sender: &Sender<CommitmentAggregationData> )

fn reset_poh_recorder( my_pubkey: &Pubkey, blockstore: &Blockstore, bank: &Arc<Bank>, poh_recorder: &RwLock<PohRecorder>, leader_schedule_cache: &LeaderScheduleCache )

fn replay_active_banks_concurrently( blockstore: &Blockstore, bank_forks: &RwLock<BankForks>, my_pubkey: &Pubkey, vote_account: &Pubkey, progress: &mut ProgressMap, transaction_status_sender: Option<&TransactionStatusSender>, verify_recyclers: &VerifyRecyclers, replay_vote_sender: &ReplayVoteSender, replay_timing: &mut ReplayTiming, log_messages_bytes_limit: Option<usize>, active_bank_slots: &[Slot], prioritization_fee_cache: &PrioritizationFeeCache ) -> Vec<ReplaySlotFromBlockstore>

fn replay_active_bank( blockstore: &Blockstore, bank_forks: &RwLock<BankForks>, my_pubkey: &Pubkey, vote_account: &Pubkey, progress: &mut ProgressMap, transaction_status_sender: Option<&TransactionStatusSender>, verify_recyclers: &VerifyRecyclers, replay_vote_sender: &ReplayVoteSender, replay_timing: &mut ReplayTiming, log_messages_bytes_limit: Option<usize>, bank_slot: Slot, prioritization_fee_cache: &PrioritizationFeeCache ) -> ReplaySlotFromBlockstore

fn process_replay_results( blockstore: &Blockstore, bank_forks: &RwLock<BankForks>, progress: &mut ProgressMap, transaction_status_sender: Option<&TransactionStatusSender>, cache_block_meta_sender: Option<&CacheBlockMetaSender>, heaviest_subtree_fork_choice: &mut HeaviestSubtreeForkChoice, bank_notification_sender: &Option<BankNotificationSender>, rewards_recorder_sender: &Option<RewardsRecorderSender>, rpc_subscriptions: &Arc<RpcSubscriptions>, duplicate_slots_tracker: &mut BTreeSet<Slot>, gossip_duplicate_confirmed_slots: &BTreeMap<Slot, Hash>, epoch_slots_frozen_slots: &mut BTreeMap<Slot, Hash>, unfrozen_gossip_verified_vote_hashes: &mut UnfrozenGossipVerifiedVoteHashes, latest_validator_votes_for_frozen_banks: &mut LatestValidatorVotesForFrozenBanks, cluster_slots_update_sender: &ClusterSlotsUpdateSender, cost_update_sender: &Sender<CostUpdate>, duplicate_slots_to_repair: &mut HashMap<Slot, Hash>, ancestor_hashes_replay_update_sender: &AncestorHashesReplayUpdateSender, block_metadata_notifier: Option<BlockMetadataNotifierLock>, replay_result_vec: &[ReplaySlotFromBlockstore], prioritization_fee_cache: &PrioritizationFeeCache, purge_repair_slot_counter: &mut BTreeMap<Slot, usize> ) -> bool

fn replay_active_banks( blockstore: &Blockstore, bank_forks: &RwLock<BankForks>, my_pubkey: &Pubkey, vote_account: &Pubkey, progress: &mut ProgressMap, transaction_status_sender: Option<&TransactionStatusSender>, cache_block_meta_sender: Option<&CacheBlockMetaSender>, verify_recyclers: &VerifyRecyclers, heaviest_subtree_fork_choice: &mut HeaviestSubtreeForkChoice, replay_vote_sender: &ReplayVoteSender, bank_notification_sender: &Option<BankNotificationSender>, rewards_recorder_sender: &Option<RewardsRecorderSender>, rpc_subscriptions: &Arc<RpcSubscriptions>, duplicate_slots_tracker: &mut BTreeSet<Slot>, gossip_duplicate_confirmed_slots: &BTreeMap<Slot, Hash>, epoch_slots_frozen_slots: &mut BTreeMap<Slot, Hash>, unfrozen_gossip_verified_vote_hashes: &mut UnfrozenGossipVerifiedVoteHashes, latest_validator_votes_for_frozen_banks: &mut LatestValidatorVotesForFrozenBanks, cluster_slots_update_sender: &ClusterSlotsUpdateSender, cost_update_sender: &Sender<CostUpdate>, duplicate_slots_to_repair: &mut HashMap<Slot, Hash>, ancestor_hashes_replay_update_sender: &AncestorHashesReplayUpdateSender, block_metadata_notifier: Option<BlockMetadataNotifierLock>, replay_timing: &mut ReplayTiming, log_messages_bytes_limit: Option<usize>, replay_slots_concurrently: bool, prioritization_fee_cache: &PrioritizationFeeCache, purge_repair_slot_counter: &mut BTreeMap<Slot, usize> ) -> bool

pub fn compute_bank_stats( my_vote_pubkey: &Pubkey, ancestors: &HashMap<u64, HashSet<u64>>, frozen_banks: &mut Vec<Arc<Bank>>, tower: &mut Tower, progress: &mut ProgressMap, vote_tracker: &VoteTracker, cluster_slots: &ClusterSlots, bank_forks: &RwLock<BankForks>, heaviest_subtree_fork_choice: &mut HeaviestSubtreeForkChoice, latest_validator_votes_for_frozen_banks: &mut LatestValidatorVotesForFrozenBanks ) -> Vec<Slot>

fn update_propagation_status( progress: &mut ProgressMap, slot: Slot, bank_forks: &RwLock<BankForks>, vote_tracker: &VoteTracker, cluster_slots: &ClusterSlots )

pub fn select_vote_and_reset_forks( heaviest_bank: &Arc<Bank>, heaviest_bank_on_same_voted_fork: Option<&Arc<Bank>>, ancestors: &HashMap<u64, HashSet<u64>>, descendants: &HashMap<u64, HashSet<u64>>, progress: &ProgressMap, tower: &mut Tower, latest_validator_votes_for_frozen_banks: &LatestValidatorVotesForFrozenBanks, fork_choice: &HeaviestSubtreeForkChoice ) -> SelectVoteAndResetForkResult

Given a heaviest_bank and a heaviest_bank_on_same_voted_fork, return a bank to vote on, a bank to reset to, and a list of switch failure reasons.

If heaviest_bank_on_same_voted_fork is None due to that fork no longer being valid to vote on, it’s possible that a validator will not be able to reset away from the invalid fork that they last voted on. To resolve this scenario, validators need to wait until they can create a switch proof for another fork or until the invalid fork is be marked valid again if it was confirmed by the cluster. Until this is resolved, leaders will build each of their blocks from the last reset bank on the invalid fork.

fn update_fork_propagated_threshold_from_votes( progress: &mut ProgressMap, newly_voted_pubkeys: Vec<Pubkey>, cluster_slot_pubkeys: Vec<Pubkey>, fork_tip: Slot, bank_forks: &BankForks )

fn update_slot_propagated_threshold_from_votes( newly_voted_pubkeys: &mut Vec<Pubkey>, cluster_slot_pubkeys: &mut Vec<Pubkey>, leader_bank: &Bank, leader_propagated_stats: &mut PropagatedStats, did_child_reach_threshold: bool ) -> bool

fn mark_slots_confirmed( confirmed_forks: &[(Slot, Hash)], blockstore: &Blockstore, bank_forks: &RwLock<BankForks>, progress: &mut ProgressMap, duplicate_slots_tracker: &mut BTreeSet<Slot>, fork_choice: &mut HeaviestSubtreeForkChoice, epoch_slots_frozen_slots: &mut BTreeMap<Slot, Hash>, duplicate_slots_to_repair: &mut HashMap<Slot, Hash>, ancestor_hashes_replay_update_sender: &AncestorHashesReplayUpdateSender, purge_repair_slot_counter: &mut BTreeMap<Slot, usize> )

fn confirm_forks( tower: &Tower, voted_stakes: &VotedStakes, total_stake: Stake, progress: &ProgressMap, bank_forks: &RwLock<BankForks> ) -> Vec<(Slot, Hash)>

pub fn handle_new_root( new_root: Slot, bank_forks: &RwLock<BankForks>, progress: &mut ProgressMap, accounts_background_request_sender: &AbsRequestSender, highest_confirmed_root: Option<Slot>, heaviest_subtree_fork_choice: &mut HeaviestSubtreeForkChoice, duplicate_slots_tracker: &mut BTreeSet<Slot>, gossip_duplicate_confirmed_slots: &mut BTreeMap<Slot, Hash>, unfrozen_gossip_verified_vote_hashes: &mut UnfrozenGossipVerifiedVoteHashes, has_new_vote_been_rooted: &mut bool, voted_signatures: &mut Vec<Signature>, epoch_slots_frozen_slots: &mut BTreeMap<Slot, Hash>, drop_bank_sender: &Sender<Vec<Arc<Bank>>> )

fn generate_new_bank_forks( blockstore: &Blockstore, bank_forks: &RwLock<BankForks>, leader_schedule_cache: &Arc<LeaderScheduleCache>, rpc_subscriptions: &Arc<RpcSubscriptions>, progress: &mut ProgressMap, replay_timing: &mut ReplayTiming )

fn new_bank_from_parent_with_notify( parent: &Arc<Bank>, slot: u64, root_slot: u64, leader: &Pubkey, rpc_subscriptions: &Arc<RpcSubscriptions>, new_bank_options: NewBankOptions ) -> Bank

fn record_rewards( bank: &Bank, rewards_recorder_sender: &Option<RewardsRecorderSender> )

pub fn get_unlock_switch_vote_slot(cluster_type: ClusterType) -> Slot

pub fn join(self) -> Result<()>