Struct solana_core::unprocessed_transaction_storage::ThreadLocalUnprocessedPackets

pub struct ThreadLocalUnprocessedPackets {
    unprocessed_packet_batches: UnprocessedPacketBatches,
    thread_type: ThreadType,
}

Fields
unprocessed_packet_batches: UnprocessedPacketBatchesthread_type: ThreadType
Implementations

impl ThreadLocalUnprocessedPackets

fn is_empty(&self) -> bool

pub fn thread_type(&self) -> ThreadType

fn len(&self) -> usize

fn max_receive_size(&self) -> usize

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut DeserializedPacket>

fn forward_option(&self) -> ForwardOption

fn clear(&mut self)

fn insert_batch( &mut self, deserialized_packets: Vec<ImmutableDeserializedPacket> ) -> PacketBatchInsertionMetrics

fn filter_forwardable_packets_and_add_batches( &mut self, bank: Arc<Bank>, forward_buffer: &mut ForwardPacketBatchesByAccounts ) -> FilterForwardingResults

Filter out packets that fail to sanitize, or are no longer valid (could be too old, a duplicate of something already processed). Doing this in batches to avoid checking bank’s blockhash and status cache per transaction which could be bad for performance. Added valid and sanitized packets to forwarding queue.

fn take_priority_queue( &mut self ) -> MinMaxHeap<Arc<ImmutableDeserializedPacket>>

Take self.unprocessed_packet_batches’s priority_queue out, leave empty MinMaxHeap in its place.

fn verify_priority_queue(&self, original_capacity: usize)

Verify that the priority queue and map are consistent and that original capacity is maintained.

fn sanitize_unforwarded_packets( &mut self, packets_to_process: &[Arc<ImmutableDeserializedPacket>], bank: &Arc<Bank> ) -> (Vec<SanitizedTransaction>, Vec<usize>)

sanitize un-forwarded packet into SanitizedTransaction for validation and forwarding.

fn filter_invalid_transactions( transactions: &[SanitizedTransaction], bank: &Arc<Bank> ) -> Vec<usize>

Checks sanitized transactions against bank, returns valid transaction indexes

fn prepare_filtered_packet_indexes( transaction_to_packet_indexes: &[usize], retained_transaction_indexes: &[usize] ) -> Vec<usize>

fn add_filtered_packets_to_forward_buffer( forward_buffer: &mut ForwardPacketBatchesByAccounts, packets_to_process: &[Arc<ImmutableDeserializedPacket>], transactions: &[SanitizedTransaction], transaction_to_packet_indexes: &[usize], forwardable_transaction_indexes: &[usize], dropped_tx_before_forwarding_count: &mut usize, feature_set: &FeatureSet ) -> Vec<usize>

try to add filtered forwardable and valid packets to forward buffer; returns vector of packet indexes that were accepted for forwarding.

fn collect_retained_packets( message_hash_to_transaction: &mut HashMap<Hash, DeserializedPacket>, packets_to_process: &[Arc<ImmutableDeserializedPacket>], retained_packet_indexes: &[usize] ) -> Vec<Arc<ImmutableDeserializedPacket>>

fn remove_non_retained_packets( message_hash_to_transaction: &mut HashMap<Hash, DeserializedPacket>, packets_to_process: &[Arc<ImmutableDeserializedPacket>], retained_packet_indexes: &[usize] )

remove packets from UnprocessedPacketBatches.message_hash_to_transaction after they have been removed from UnprocessedPacketBatches.packet_priority_queue

fn process_packets<F>( &mut self, bank: &Bank, banking_stage_stats: &BankingStageStats, slot_metrics_tracker: &mut LeaderSlotMetricsTracker, processing_function: F ) -> boolwhere F: FnMut(&Vec<Arc<ImmutableDeserializedPacket>>, &mut ConsumeScannerPayload<'_>) -> Option<Vec<usize>>,

fn prepare_packets_to_forward( &self, packets_to_forward: impl Iterator<Item = Arc<ImmutableDeserializedPacket>>, total_tracer_packets_in_buffer: &mut usize ) -> (Vec<Arc<ImmutableDeserializedPacket>>, Vec<Arc<ImmutableDeserializedPacket>>, Vec<bool>)

Prepare a chunk of packets for forwarding, filter out already forwarded packets while counting tracers. Returns Vec of unforwarded packets, and Vec of same size each indicates corresponding packet is tracer packet.